Made required package description slightly more accurate.
I haven't verified that the packages I recommend are a "smallest set"
for any meaningful purpose. But they are somewhat small.
I might try to make them smaller, but not tonight.

Now it is time for honk-shoo mimimimimi.

amendment: I now have good reason to believe that the set of packages I
recommend is the minimal set of packages, as in takes the least time to
install, assuming the user is only willing to install debian packages.
[The documentation for `texlive-fonts-extra` for Debian 13.0
(trixie)](https://packages.debian.org/trixie/texlive-fonts-extra)
makes no mention of any dependencies that provide only the fonts we
need, and [the documentation for that package on
sid](https://packages.debian.org/sid/texlive-fonts-extra) explicitly says the
fonts we need are in a CTAN package that is a *member* of this debian
package - so, I think, not separately installable with `apt`.